refactor: pinch hook 분리
